% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generer_rapport.R
\name{generer_rapport}
\alias{generer_rapport}
\title{Générer un rapport d'analyse}
\usage{
generer_rapport(commune, departement, output)
}
\arguments{
\item{commune}{Nom de la commune.}

\item{departement}{Nom du département.}

\item{output}{Chemin de sortie du rapport.}
}
\value{
Génère un rapport HTML.
}
\description{
Cette fonction compile un rapport Quarto avec des paramètres spécifiques.
}
